move custom assertions into shared setup
